# openapi/buscai-v2.yaml
openapi: 3.1.0
info:
  title: BUSCAI API
  version: 1.0.0
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Niche:
      type: object
      required: [id, label, slug]
      properties:
        id: { type: string }
        label: { type: string }
        slug: { type: string }
        isActive: { type: boolean }
...
paths:
...
  /niches:
    get:
      summary: List niches
      operationId: listNiches
      tags: [catalog]
      parameters:
        - name: cityId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of niches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Niche"
...